Compression module support

+++ Runtime/ModManager.cs
@@ -99,29 +99,20 @@ namespace ModIO
                 }
                 else
                 {
-
                     DataStorage.DeleteDirectory(tempLocation, (dd_path, dd_success) =>
                     DataStorage.CreateDirectory(tempLocation, (cd_path, cd_success) =>
                     {
                         if(dd_success && cd_success)
                         {
-                            // extract
-                            try
-                            {
-                                using (var zip = Ionic.Zip.ZipFile.Read(archivePath))
-                                {
-                                    zip.ExtractAll(tempLocation);
-                                }
+                            bool didExtract = CompressionModule.ExtractAll(archivePath, tempLocation);
 
+                            if(didExtract)
+                            {
                                 // Remove old versions
                                 ModManager.UninstallMod(modId, onOldVersionsUninstalled);
                             }
-                            catch(Exception e)
+                            else
                             {
-                                Debug.LogWarning("[mod.io] Unable to extract binary to a temporary folder."
-                                                 + "\nLocation: " + tempLocation + "\n\n"
-                                                 + Utility.GenerateExceptionDebugString(e));
-
                                 DataStorage.DeleteDirectory(tempLocation, null);
 
                                 if(onComplete != null)
@@ -1521,46 +1512,20 @@ namespace ModIO
                 {
                     if(success)
                     {
-                        try
+                        success = CompressionModule.CompressFileCollection(rootDirectory, fileList, archiveFilePath);
+
+                        if(success)
                         {
-                            using(var zip = new Ionic.Zip.ZipFile())
-                            {
-                                foreach(string filePath in fileList)
-                                {
-                                    string relativeFilePath = filePath.Substring(rootDirectoryLength);
-                                    string relativeDirectory = Path.GetDirectoryName(relativeFilePath);
-
-                                    zip.AddFile(filePath, relativeDirectory);
-                                    zip.Save(archiveFilePath);
-                                }
-                            }
-
                             UploadModBinary_Zipped(modId, modfileValues, archiveFilePath, setActiveBuild, onSuccess, onError);
                         }
-                        catch(Exception e)
-                        {
-                            Debug.LogError("[mod.io] Unable to zip mod binary prior to uploading.\n\n"
-                                           + Utility.GenerateExceptionDebugString(e));
-
-                            if(onError != null)
-                            {
-                                WebRequestError error = WebRequestError.GenerateLocal("Unable to zip mod binary prior to uploading");
-
-                                onError(error);
-                            }
-                        }
                     }
-                    else
+
+                    if(!success)
                     {
-                        string errorString = ("Unable to create directory for mod binary archive prior to upload."
-                                              + "\nDirectory: " + path);
-
-                        Debug.LogError("[mod.io] " + errorString);
-
                         if(onError != null)
                         {
-                            WebRequestError error = WebRequestError.GenerateLocal(errorString);
-                            onError(error);
+                            WebRequestError error = WebRequestError.GenerateLocal("Unable to zip mod binary prior to uploading");
+                            onError.Invoke(error);
                         }
                     }
                 });
@@ -1582,32 +1547,19 @@ namespace ModIO
 
             DataStorage.CreateDirectory(Path.GetDirectoryName(binaryZipLocation), (path, success) =>
             {
-                try
-                {
-                    using(var zip = new Ionic.Zip.ZipFile())
-                    {
-                        zip.AddFile(unzippedBinaryLocation, "");
-                        zip.Save(binaryZipLocation);
-                    }
-
-                    zipSucceeded = true;
-                }
-                catch(Exception e)
-                {
-                    Debug.LogError("[mod.io] Unable to zip mod binary prior to uploading.\n\n"
-                                   + Utility.GenerateExceptionDebugString(e));
-
-                    if(onError != null)
-                    {
-                        WebRequestError error = WebRequestError.GenerateLocal("Unable to zip mod binary prior to uploading");
-
-                        onError(error);
-                    }
-                }
+                zipSucceeded = CompressionModule.CompressFile(unzippedBinaryLocation, binaryZipLocation);
 
                 if(zipSucceeded)
                 {
                     UploadModBinary_Zipped(modId, modfileValues, binaryZipLocation, setActiveBuild, onSuccess, onError);
+                }
+                else
+                {
+                    if(onError != null)
+                    {
+                        WebRequestError error = WebRequestError.GenerateLocal("Unable to zip mod binary prior to uploading");
+                        onError.Invoke(error);
+                    }
                 }
             });
         }
